fixed chatroom
